Separate precision evaluator
